services/modeling/pyproject.toml
[project]
name = "fightmatch-modeling"
version = "0.1.0"
description = "Win-probability model + matchmaking engine for FightMatch"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "joblib>=1.3.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
fightmatch-train = "modeling.cli:train_cmd"
fightmatch-recommend = "modeling.cli:recommend_cmd"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
